Added Word 2003 images and image title Some commenting and cleanup of code

+++ src/fi/kaila/suku/report/XmlReport.java
@@ -7,6 +7,7 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Formatter;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.prefs.Preferences;
@@ -59,6 +60,7 @@ public class XmlReport implements ReportInterface {
 	private String report = null;
 	private String folder = null;
 	private int maxImageWidth = 0;
+	private int translatorIdx;
 	private Logger logger = Logger.getLogger(this.getClass().getName());
 
 	int imageCounter = 0;
@@ -71,6 +73,7 @@ public class XmlReport implements ReportInterface {
 			throws SukuException {
 		this.parent = parent;
 		this.title = title;
+		this.translatorIdx = translatorIdx;
 		maxImageWidth = parent.getImageMaxWidth();
 		switch (translatorIdx) {
 		case 1:
@@ -103,75 +106,78 @@ public class XmlReport implements ReportInterface {
 				f.delete();
 
 			}
-			int lastIdx = report.replace('\\', '/').lastIndexOf('.');
-			if (lastIdx > 0) {
-				folder = report.substring(0, lastIdx) + "_files";
-				File d = new File(folder);
-				if (d.exists()) {
-					if (!fileExists && d.isDirectory()) {
-						int resu = JOptionPane
-								.showConfirmDialog(
-										parent,
-										Resurses
-												.getString("CONFIRM_REPLACE_REPORTDIR"),
-										Resurses.getString(Resurses.SUKU),
-										JOptionPane.YES_NO_OPTION,
-										JOptionPane.QUESTION_MESSAGE);
-						if (resu != JOptionPane.YES_OPTION) {
-							throw new SukuException("");
-						}
-						fileExists = true;
-					}
-					if (fileExists && d.isDirectory()) {
-
-						String[] files = d.list();
-
-						for (int i = 0; i < files.length; i++) {
-							File df = new File(folder + "/" + files[i]);
-							df.delete();
-						}
-
-					}
-				}
-				d.mkdirs();
-				//
-				// css-file is currently embedded in html-file
-				// and separate css-file is not needed
-				//
-				if (translatorIdx == 2 && false) {
-
-					InputStream in = this.getClass().getResourceAsStream(
-							"/xml/finfamily.css");
-
-					BufferedInputStream bis = new BufferedInputStream(in);
-
-					File ff = new File(folder + "/finfamily.css");
-					byte[] buffi = new byte[1024];
-					FileOutputStream fos;
-
-					try {
-						fos = new FileOutputStream(ff);
-
-						while (true) {
-
-							int retu = in.read(buffi);
-							if (retu == -1) {
-								break;
+			if (translatorIdx > 1) { // word 2003 format does not use folder
+				int lastIdx = report.replace('\\', '/').lastIndexOf('.');
+				if (lastIdx > 0) {
+					folder = report.substring(0, lastIdx) + "_files";
+					File d = new File(folder);
+					if (d.exists()) {
+						if (!fileExists && d.isDirectory()) {
+							int resu = JOptionPane
+									.showConfirmDialog(
+											parent,
+											Resurses
+													.getString("CONFIRM_REPLACE_REPORTDIR"),
+											Resurses.getString(Resurses.SUKU),
+											JOptionPane.YES_NO_OPTION,
+											JOptionPane.QUESTION_MESSAGE);
+							if (resu != JOptionPane.YES_OPTION) {
+								throw new SukuException("");
 							}
-							fos.write(buffi, 0, retu);
+							fileExists = true;
+						}
+						if (fileExists && d.isDirectory()) {
+
+							String[] files = d.list();
+
+							for (int i = 0; i < files.length; i++) {
+								File df = new File(folder + "/" + files[i]);
+								df.delete();
+							}
 
 						}
-					} catch (FileNotFoundException e) {
-						logger.log(Level.WARNING, "css-file", e);
-					} catch (IOException e) {
-						logger.log(Level.WARNING, "css-file", e);
 					}
+					d.mkdirs();
+				}
+			}
+			//
+			// css-file is currently embedded in html-file
+			// and separate css-file is not needed
+			//
+			if (translatorIdx == 2 && false) {
 
+				InputStream in = this.getClass().getResourceAsStream(
+						"/xml/finfamily.css");
+
+				BufferedInputStream bis = new BufferedInputStream(in);
+
+				File ff = new File(folder + "/finfamily.css");
+				byte[] buffi = new byte[1024];
+				FileOutputStream fos;
+
+				try {
+					fos = new FileOutputStream(ff);
+
+					while (true) {
+
+						int retu = in.read(buffi);
+						if (retu == -1) {
+							break;
+						}
+						fos.write(buffi, 0, retu);
+
+					}
+				} catch (FileNotFoundException e) {
+					logger.log(Level.WARNING, "css-file", e);
+				} catch (IOException e) {
+					logger.log(Level.WARNING, "css-file", e);
 				}
 
-				return;
 			}
+
+			return;
 		}
+
 		throw new SukuException(Resurses.getString("WARN_REPORT_NOT_SELECTED"));
 
 	}
@@ -226,15 +232,16 @@ public class XmlReport implements ReportInterface {
 	public void addText(BodyText bt) {
 		Element ele;
 		Element tele;
-
+		Element iii;
 		String imgName = null;
 		String imgTitle = null;
 		int imgWidth = 0;
-
+		int imgHeight = 0;
+		String img = null;
 		if (bt instanceof ImageText) {
 			ImageText it = (ImageText) bt;
 			imgWidth = it.getWidth();
-
+			imgHeight = it.getHeight();
 			imgTitle = it.getImageTitle();
 
 			imageCounter++;
@@ -243,9 +250,16 @@ public class XmlReport implements ReportInterface {
 
 			FileOutputStream fos;
 			try {
-				fos = new FileOutputStream(ff);
-				fos.write(it.getData());
-				fos.close();
+
+				if (translatorIdx == 1) {
+					img = convertTo64(it);
+
+				} else {
+					fos = new FileOutputStream(ff);
+					fos.write(it.getData());
+					fos.close();
+				}
+
 			} catch (FileNotFoundException e) {
 				logger.log(Level.WARNING, "Image", e);
 			} catch (IOException e) {
@@ -267,11 +281,27 @@ public class XmlReport implements ReportInterface {
 		if (imgName != null) {
 			ele.setAttribute("image", imgName);
 			ele.setAttribute("title", imgTitle);
+			ele.setAttribute("imageNo", "" + imageCounter);
+			ele.setAttribute("imageName", Resurses.getString("REPORT.IMAGE"));
 			if (imgWidth > maxImageWidth) {
-				imgWidth = maxImageWidth;
-			}
-			if (imgWidth > 10) {
+				float mw = maxImageWidth;
+
+				float multip = mw / imgWidth;
+				float w = imgWidth * multip;
+				float h = imgHeight * multip;
+
+				if (w > 10) {
+					ele.setAttribute("width", "" + w);
+					ele.setAttribute("height", "" + h);
+				}
+			} else {
 				ele.setAttribute("width", "" + imgWidth);
+				ele.setAttribute("height", "" + imgHeight);
+			}
+			if (img != null) {
+				iii = doc.createElement("media");
+				iii.setTextContent(img);
+				ele.appendChild(iii);
 			}
 		}
 		String prevStyle = "";
@@ -333,12 +363,11 @@ public class XmlReport implements ReportInterface {
 			// delete folder if empty
 			// note that delete does not delete it if it's not empty
 			//
-			File dd = new File(folder);
-			dd.delete();
-
+			if (folder != null) {
+				File dd = new File(folder);
+				dd.delete();
+			}
 			if (translator != null) {
-				// System.setProperty("javax.xml.transform.TransformerFactory",
-				// "net.sf.saxon.TransformerFactoryImpl");
 
 				DOMSource docw = new DOMSource(doc);
 				ByteArrayOutputStream bout = new ByteArrayOutputStream();
@@ -372,16 +401,11 @@ public class XmlReport implements ReportInterface {
 				}
 
 			} else {
-				// // Transform the source XML to System.out.
-				// transformer.transform(new StreamSource(sourceID),
-				// new StreamResult(System.out));
-				//		        
 
 				DOMSource docw = new DOMSource(doc);
 				ByteArrayOutputStream bout = new ByteArrayOutputStream();
 				TransformerFactory tfactory = TransformerFactory.newInstance();
-				// Templates templates = tfactory.newTemplates(new
-				// StreamSource(xslt));
+
 				Transformer transformer = tfactory.newTransformer();//
 				transformer.transform(docw, new StreamResult(bout));
 				FileOutputStream fos = new FileOutputStream(report);
@@ -456,6 +480,92 @@ public class XmlReport implements ReportInterface {
 			String messu = e.getMessage();
 			throw new SukuException(messu);
 		}
+	}
+
+	static final String table64 = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
+
+	/**
+	 * Copied from XmlReports::CopyBase64File
+	 * 
+	 * @param it
+	 *            the Style with the image
+	 * @param to
+	 *            path for output file
+	 */
+	private String convertTo64(ImageText it) {
+
+		FileOutputStream fos;
+		Formatter ff = new Formatter();
+		float fw = it.getWidth();
+		float fh = it.getHeight();
+		StringBuilder sbb = new StringBuilder();
+		// ff.format("<image width=\"%1.2f\" height=\"%1.2f\">\r\n", fw, fh);
+
+		byte[] ibuf = new byte[3];
+		byte[] obuf = new byte[4];
+
+		int inputparts;
+		int colPos = 0;
+		int insize = it.getData().length;
+		int indata = 0;
+
+		// char *indata = (char *)buffer;
+
+		while (insize > 0) {
+			for (int i = inputparts = 0; i < 3; i++) {
+				if (insize > 0) {
+					inputparts++;
+
+					ibuf[i] = it.getData()[indata];
+					indata++;
+					insize--;
+				} else
+					ibuf[i] = 0;
+			}
+
+			obuf[0] = (byte) ((ibuf[0] & 0xFC) >> 2);
+			obuf[1] = (byte) (((ibuf[0] & 0x03) << 4) | ((ibuf[1] & 0xF0) >> 4));
+			obuf[2] = (byte) (((ibuf[1] & 0x0F) << 2) | ((ibuf[2] & 0xC0) >> 6));
+			obuf[3] = (byte) (ibuf[2] & 0x3F);
+
+			char c0, c1, c2, c3;
+			switch (inputparts) {
+			case 1: /* only one byte read */
+				c0 = table64.charAt(obuf[0]);
+				c1 = table64.charAt(obuf[1]);
+				ff.format("%c%c==", c0, c1);
+
+				break;
+			case 2: /* two bytes read */
+				c0 = table64.charAt(obuf[0]);
+				c1 = table64.charAt(obuf[1]);
+				c2 = table64.charAt(obuf[2]);
+				ff.format("%c%c%c=", c0, c1, c2);
+				break;
+			default:
+				c0 = table64.charAt(obuf[0]);
+				c1 = table64.charAt(obuf[1]);
+				c2 = table64.charAt(obuf[2]);
+				c3 = table64.charAt(obuf[3]);
+				ff.format("%c%c%c%c", c0, c1, c2, c3);
+				break;
+			}
+
+			if (colPos >= 76) {
+				ff.format("\n");
+
+				colPos = 0;
+			} else {
+				colPos += 4;
+			}
+
+		}
+
+		// ff.format("</image>\r\n");
+		sbb.append(ff.out());
+
+		return sbb.toString();
+
 	}
 
 }
